+++ Source/DesignSystem/index.ts
@@ -18,7 +18,6 @@ export { Icon, IconProps } from './atoms/Icon';
 export { IconButton, IconButtonProps } from './atoms/IconButton';
 export { Link, LinkProps } from './atoms/Link';
 export { FullPageLoadingSpinner, LoadingSpinner } from './atoms/LoadingSpinner';
-export { MaintenanceMessageBuilding } from './atoms/MaintenanceMessage';
 export { StatusIndicator, StatusIndicatorProps } from './atoms/StatusIndicator';
 export { Summary } from './atoms/Metrics';
 export { Tabs } from './atoms/Tabs';
@@ -34,6 +33,7 @@ export { FileUploadForm, FileUploadFormProps, FileUploadFormRef } from './molecu
 export { Graph } from './molecules/Metrics/Graph';
 export { DropdownMenu, DropdownMenuProps, MenuItemProps } from './molecules/DropdownMenu';
 export { MenuList, MenuListProps } from './molecules/MenuList';
+export { ProblemPage } from './molecules/ProblemPage';
 export { NavigationBar, NavigationBarProps, getPrimaryNavigationItems, getSecondaryNavigationItems, getSelectionMenuItems } from './molecules/NavigationBar';
 export { SelectCard, SimpleCard, SimpleCardProps, SimpleCardGrid, SimpleCardGridProps } from './molecules/Card';
 export { SidePanel, SidePanelProps, getSidePanelItems } from './molecules/SidePanel';