+++ Source/SelfService/Web/apis/solutions/application.ts
@@ -19,8 +19,6 @@ export type HttpApplicationEnvironment = {
 export type HttpApplicationRequest = {
     id: string;
     name: string;
-    //contactName: string;
-    //contactEmail: string;
     environments: HttpApplicationEnvironment[];
 };
 
@@ -109,10 +107,12 @@ export async function isApplicationOnline(applicationID: string): Promise<Applic
             mode: 'cors',
         });
 
-    return response;
+    const data: ApplicationBuildState = await parseJSONResponse(response);
+
+    return data;
 };
 
-export async function getLiveApplications(): Promise<any> {
+export async function getLiveApplications(): Promise<HttpResponseApplications> {
     const url = `${getServerUrlPrefix()}/live/applications`;
 
     const result = await fetch(
@@ -122,7 +122,8 @@ export async function getLiveApplications(): Promise<any> {
             mode: 'cors',
         });
 
-    const jsonResult = await result.json();
+    const jsonResult: HttpResponseApplications = await result.json();
+    jsonResult.applications = jsonResult.applications || [];
     return jsonResult;
 };