+++ Source/SelfService/Web/CHANGELOG.md
@@ -1,3 +1,26 @@
+# [3.11.1] - 2023-8-30 [PR: #463](https://github.com/dolittle/Studio/pull/463)
+## Summary
+
+Improve creating new application with better navigation and application building page.
+
+### Added
+
+- ProblemPage to Design System
+
+
+# [3.11.0] - 2023-8-30 [PR: #462](https://github.com/dolittle/Studio/pull/462)
+- do not disable authentication type based on connector deployment type
+- Simplify selection of configuration sections. MDP will now always be visible until we have an M3 hosting type added
+- Slighty improve the check to evaluate if basicauth has any values set (ie: is in use)
+
+
+# [3.10.0] - 2023-8-29 [PR: #461](https://github.com/dolittle/Studio/pull/461)
+## Summary
+- Support for selecting authentication type when setting up the M3 connector
+- Cloud connector supports only ION
+- On premises connector supports MDP/ION and MDP/M3 basic authentication
+
+
 # [3.9.15] - 2023-8-28 [PR: #460](https://github.com/dolittle/Studio/pull/460)
 ## Summary